2020 - Semestre 2/Análise e Desenvolvimento de Sistemas/5°/(Equipe 3) LGPD-compliant-website/views/index.py
from views.account import Signin
from flask.helpers import url_for
from werkzeug.utils import redirect
from .blueprint import app
 

@app.route('/')
def index():
    return redirect(url_for(Signin.ROUTE), code=302)