2020 - Semestre 2/Análise e Desenvolvimento de Sistemas/5°/(Equipe 3) LGPD-compliant-website/requirements.txt
azure-common==1.1.26
azure-core==1.9.0
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
Babel==2.8.0
cryptography==3.2.1
debugpy==1.0.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
flask-talisman==0.7.0
Flask-WTF==0.14.3
gunicorn==20.0.4
Jinja2==2.11.2
passlib==1.7.4
phonenumbers==8.12.11
psycopg2-binary==2.8.6
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.4